+++ core-bundle/src/Resources/contao/library/Contao/Model.php
@@ -394,9 +394,9 @@ abstract class Model
 	 */
 	public function markModified($strKey)
 	{
-		if (!isset($this->arrModified[$strKey]) && isset($this->arrData[$strKey]))
+		if (!isset($this->arrModified[$strKey]))
 		{
-			$this->arrModified[$strKey] = $this->arrData[$strKey];
+			$this->arrModified[$strKey] = $this->arrData[$strKey] ?? null;
 		}
 	}