+++ core-bundle/tests/DependencyInjection/ContaoCoreExtensionTest.php
@@ -36,6 +36,7 @@ use Contao\CoreBundle\EventListener\MergeHttpHeadersListener;
 use Contao\CoreBundle\EventListener\PrettyErrorScreenListener;
 use Contao\CoreBundle\EventListener\RefererIdListener;
 use Contao\CoreBundle\EventListener\ResponseExceptionListener;
+use Contao\CoreBundle\EventListener\SessionListener;
 use Contao\CoreBundle\EventListener\StoreRefererListener;
 use Contao\CoreBundle\EventListener\ToggleViewListener;
 use Contao\CoreBundle\EventListener\UserSessionListener;
@@ -400,6 +401,22 @@ class ContaoCoreExtensionTest extends TestCase
         $this->assertSame('kernel.exception', $tags['kernel.event_listener'][0]['event']);
         $this->assertSame('onKernelException', $tags['kernel.event_listener'][0]['method']);
         $this->assertSame(64, $tags['kernel.event_listener'][0]['priority']);
+    }
+
+    /**
+     * Tests the contao.listener.session_listener service.
+     */
+    public function testRegistersTheSessionListener()
+    {
+        $this->assertTrue($this->container->has('contao.listener.session_listener'));
+
+        $definition = $this->container->getDefinition('contao.listener.session_listener');
+
+        $this->assertSame(SessionListener::class, $definition->getClass());
+        $this->assertSame('session_listener', $definition->getDecoratedService()[0]);
+        $this->assertSame('contao.listener.session_listener.inner', (string) $definition->getArgument(0));
+        $this->assertSame('contao.framework', (string) $definition->getArgument(1));
+        $this->assertSame('contao.routing.scope_matcher', (string) $definition->getArgument(2));
     }
 
     /**