+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Correctly track modified fields in the `Model` class (see #1290).
  * Use the normalized package versions for the CDN scripts (see #1391).
 
 ### 4.5.4 (2018-02-14)