+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+ * Do not make the response private when saving the session (see #1388).
  * Improve the folder hashing performance (see contao/core#8856).
 
 ## 4.4.14 (2018-02-14)