+++ core-bundle/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## DEV
+## 4.4.15 (2018-03-06)
 
  * Do not make the response private when saving the session (see #1388).
  * Improve the folder hashing performance (see contao/core#8856).